+++ assets/javascripts/kitten/components/cards/crowdfunding-card/test.js
@@ -150,4 +150,35 @@ describe('<CrowdfundingCard />', () => {
       expect(component).toMatchSnapshot()
     })
   })
+
+  describe('CrowdfundingCard with forceVersion prop', () => {
+    beforeEach(() => {
+      window.matchMedia = createMockMediaMatcher(false)
+
+      component = renderer
+        .create(
+          <CrowdfundingCardWidget
+            ownerTitle="Owner title"
+            ownerDescription="Owner description"
+            imageProps={{ src: '#image', alt: '' }}
+            avatarProps={{ src: '#image', alt: '' }}
+            state="Sur"
+            loading={false}
+            title="Lorem ipsum…"
+            subtitle="Sed ut perspiciatis unde omnis iste natus error Lorem"
+            titleTruncate
+            subTitleTruncate
+            buttonText="Lorem ipsum"
+            dayCounter={null}
+            buttonText="Contribute"
+            forceVersion="desktop"
+          />,
+        )
+        .toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
 })